requirements.txt
# --- CORE ---
python-dotenv
fastapi
uvicorn
requests

# --- LANGCHAIN & LANGGRAPH ---
langchain-core>=0.3.0
langchain>=0.3.0
langgraph

# --- INTELIGÊNCIA ARTIFICIAL (Drivers) ---
# Google Gemini (Versões recentes para suportar Flash 1.5/2.5)
langchain-google-genai>=2.0.0
google-generativeai>=0.8.3
google-ai-generativelanguage>=0.6.10

# OpenAI (Opcional, mas mantemos para compatibilidade)
langchain-openai>=0.2.0